+++ qt/scientific_interfaces/Inelastic/Common/RunWidget/RunView.ui
